Now it runs without errors in runtime
I hopefully fixed the indentation problem :-). I don't know where exactly the erros were but I carefully rewrote the TeavmGPStyleSheet and TeavmGPCSSRule files and now it's working. Next thing to do is to make it to work, because currently there is only a white website. So no data visible, but generated html looks good. IDK where the error is.